v14092024_2

+++ ChoiceVFileSystemBlazor/Database/Supportfiles/Proxies/Interfaces/ISupportfileEntryProxy.cs
@@ -12,7 +12,7 @@ public interface ISupportfileEntryProxy
     public Task<bool> RestoreEntryAsync(Ulid id, Ulid accessId);
 
     public Task<SupportfileFileUploadDbModel?> GetFileAsync(Ulid id);
-    public int GetMaxFileSize();
+    public long GetMaxFileSize();
     public Task<bool> RenameFileAsync(Ulid id, string newName, Ulid supportfileId, Ulid accessId);
     public Task<bool> AddFileAsync(SupportfileFileUploadDbModel file, Ulid supportfileId, Ulid accessId);
     public Task<bool> DeleteFileAsync(Ulid id, Ulid supportfileId, Ulid accessId);
